+++ noodles-bcf/src/writer/vcf_record/genotypes.rs
@@ -10,7 +10,7 @@ use noodles_vcf::{
         format::{key, Key},
         record::value::{map::Format, Map},
     },
-    record::genotypes::values::field::Value,
+    record::genotypes::values::Value,
 };
 
 use crate::{
@@ -781,7 +781,7 @@ mod tests {
     fn test_write_genotypes() -> Result<(), Box<dyn std::error::Error>> {
         use vcf::{
             header::{format::key, record::value::Map},
-            record::genotypes::values::field::Value,
+            record::genotypes::values::Value,
         };
 
         let header = vcf::Header::builder()