+++ noodles-vcf/CHANGELOG.md
@@ -19,6 +19,13 @@
 
     `ReferenceBases` no longer handles "." as missing.
 
+### Removed
+
+  * vcf/record/genotypes/values: Remove `field` module.
+
+    This removes the unused `Field` struct. The `values::field::value` module
+    is moved up as `values::value`.
+
 ## 0.26.0 - 2023-03-14
 
 ### Added